+++ source/Energinet.DataHub.MeteringPoints.Application/Create/CreateMeteringPointRuleSet.cs
@@ -12,11 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.MeteringPoints.Application.Create;
 using Energinet.DataHub.MeteringPoints.Application.Validation.Rules;
 using FluentValidation;
 
-namespace Energinet.DataHub.MeteringPoints.Application.Validation
+namespace Energinet.DataHub.MeteringPoints.Application.Create
 {
     public class CreateMeteringPointRuleSet : AbstractValidator<CreateMeteringPoint>
     {
