+++ source/Energinet.DataHub.MeteringPoints.Application/Create/CreateMeteringPointRulesHelper.cs
@@ -13,10 +13,9 @@
 // limitations under the License.
 
 using System;
-using Energinet.DataHub.MeteringPoints.Application.Create;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
 
-namespace Energinet.DataHub.MeteringPoints.Application.Validation
+namespace Energinet.DataHub.MeteringPoints.Application.Create
 {
     public static class CreateMeteringPointRulesHelper
     {
